+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.12.2 - WIP
+* Fix exception when nesting a grid with action column
 
 # 0.12.1 - 2015-05-31
 * New component: Tree